refactor(comments): extract form draft lifecycle
